Fix #61 render span graph as canvas instead of SVG - Render span graph via canvas instead of SVG - Zoom range changed to [0, 1], e.g. time and trace agnostic allowed   removal of some utils - "Timeline" -> 0ms - Use props instead of context to provide span graph with viewing range - Move all span graph related classes into same folder - Misc CSS cleanup
